requirements.txt
streamlit>=1.43.0
pandas>=2.3.0
xlrd>=2.0.1
openpyxl>=3.1.5
python-dotenv>=1.0.1
supabase>=2.4.0
